fix(mobile): fence channel actions and roster mutations
